events/2025-09-27/Isham Rashik/_ToolsAgentsMCPFundamentals/tools_agents_examples/requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
python-dotenv>=1.0.0
langchain-mcp-adapters==0.1.10
langgraph>=0.1.0
fastmcp==2.12.4
mcp-use==1.3.10